Cycles: Add a new Metallic BSDF, combining condictive fresnel and multi-scattering GGX
This patch implements a new BSDF called "Metallic BSDF", which essentially is the
Glossy BSDF with Multi-Scattering GGX combined with the conductive fresnel
absorption term applied at each bounce.
To specify the properties of the conductor, either the usual 6 NK-values (real and
imaginary components of the IOR, each for the three color channels) or the
more intuitive reparametrization to "Reflectivity" and "Edge Tint" from the paper
"Artist Friendly Metallic Fresnel" can be used.

Depends on D2002.

Reviewers: #cycles, brecht

Subscribers: brecht, robocyte, harvester, swerner, jtheninja, Blendify, nutel

Differential Revision: https://developer.blender.org/D2003